test(security): isolate multi-line matcher tests; revert IsPosix dedup
The earlier IsPosix/TestPlatform dedup rewrote the SkipUnless attribute on
~11 pre-existing baselined tests across two files. Slopwatch's CI gate
scopes to files changed in a PR and flags every disabled-test (SW001) in
them, so touching those files surfaced all their pre-existing skipped tests
as new violations.

- Revert ShellApprovalMatcherTests.cs and ShellTokenizerTests.cs to their
  upstream state and drop the shared TestPlatform helper, so the PR no
  longer touches files dense with pre-existing skipped tests.
- Move the four multi-line approval-matcher tests into a dedicated
  ShellApprovalMatcherMultilineTests.cs with its own IsPosix hook.
- Baseline the four SW001 entries for the new file in
  .slopwatch/baseline.json.
